+++ package.json
@@ -11,6 +11,7 @@
     "chai": "^2.3.0",
     "del": "^1.2.0",
     "fs": "0.0.2",
+    "glob": "^5.0.14",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^2.3.1",
     "gulp-changed": "^1.2.1",
